Split gtest_coordination
